[박현민] 04-13 local 버전 프로젝트 푸쉬

+++ confirm-service/src/main/java/com/dream/confirm/dto/confirmDto.java
@@ -12,7 +12,7 @@ public class confirmDto {
 	private int orderNo;
 	private String proName;
 	private double orderPrice;
-	private int rate; // 생략
+	private double rate; // 생략
 	private String purpose;
 	private double amount;
 	private int term;
